backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.3.1
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jiter==0.11.1
numpy==2.2.6
openai==2.5.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.35
starlette==0.38.6
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
psycopg2-binary==2.9.9
pgvector==0.2.4
sentence-transformers==2.2.2
scipy==1.11.4
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
